requirements.txt
alembic==1.4.0
Babel==2.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
facebook-sdk==4.0.0rc0
Flask==1.1.1
Flask-Admin==1.5.5
Flask-BabelEx==0.9.4
Flask-Dance==3.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-OAuth==0.13
flask-paginate==0.5.5
Flask-Pagination==0.0.1
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Social==1.6.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
httplib2==0.17.0
idna==2.9
import-string==0.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonify==0.5
Mako==1.1.1
manage==0.1.13
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.18.1
oauth==1.0.1
oauth2==1.9.0.post1
oauthlib==3.1.0
pandas==1.0.1
passlib==1.7.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
query==0.1.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.13
urllib3==1.25.8
URLObject==2.4.3
validate-email==1.3
Werkzeug==1.0.0
WTForms==2.2.1
